AppEngineBigQuery/python/requirements.txt
webapp2==2.5.2
google-cloud==0.27.0
